Alpha 0.1
Added API calls to-be-made, and started drafting how all the schemas, etc. would be made. There are issues with my current MongoDB topography, so Alpha 0.2 will focus on making Items and Receipts their own collections, with Receipts and Companies pointing to Items, and Companies pointing to Receipts for later tracking.